+++ lib/docusign_esign/models/notary_recipient.rb
@@ -36,6 +36,9 @@ module DocuSign_eSign
     # 
     attr_accessor :auto_navigation
 
+    # 
+    attr_accessor :auto_responded_reason
+
     # Contains a URI for an endpoint that allows you to easily retrieve bulk recipient information.
     attr_accessor :bulk_recipients_uri
 
@@ -164,6 +167,10 @@ module DocuSign_eSign
 
     attr_accessor :phone_authentication
 
+    attr_accessor :phone_number
+
+    attr_accessor :phone_number_metadata
+
     attr_accessor :proof_file
 
     # Reserved:
@@ -276,6 +283,7 @@ module DocuSign_eSign
         :'agent_can_edit_name' => :'agentCanEditName',
         :'allow_system_override_for_locked_recipient' => :'allowSystemOverrideForLockedRecipient',
         :'auto_navigation' => :'autoNavigation',
+        :'auto_responded_reason' => :'autoRespondedReason',
         :'bulk_recipients_uri' => :'bulkRecipientsUri',
         :'can_sign_offline' => :'canSignOffline',
         :'client_user_id' => :'clientUserId',
@@ -324,6 +332,8 @@ module DocuSign_eSign
         :'note_metadata' => :'noteMetadata',
         :'offline_attributes' => :'offlineAttributes',
         :'phone_authentication' => :'phoneAuthentication',
+        :'phone_number' => :'phoneNumber',
+        :'phone_number_metadata' => :'phoneNumberMetadata',
         :'proof_file' => :'proofFile',
         :'recipient_attachments' => :'recipientAttachments',
         :'recipient_authentication_status' => :'recipientAuthenticationStatus',
@@ -375,6 +385,7 @@ module DocuSign_eSign
         :'agent_can_edit_name' => :'String',
         :'allow_system_override_for_locked_recipient' => :'String',
         :'auto_navigation' => :'String',
+        :'auto_responded_reason' => :'String',
         :'bulk_recipients_uri' => :'String',
         :'can_sign_offline' => :'String',
         :'client_user_id' => :'String',
@@ -423,6 +434,8 @@ module DocuSign_eSign
         :'note_metadata' => :'PropertyMetadata',
         :'offline_attributes' => :'OfflineAttributes',
         :'phone_authentication' => :'RecipientPhoneAuthentication',
+        :'phone_number' => :'RecipientPhoneNumber',
+        :'phone_number_metadata' => :'PropertyMetadata',
         :'proof_file' => :'RecipientProofFile',
         :'recipient_attachments' => :'Array<RecipientAttachment>',
         :'recipient_authentication_status' => :'AuthenticationStatus',
@@ -503,6 +516,10 @@ module DocuSign_eSign
 
       if attributes.has_key?(:'autoNavigation')
         self.auto_navigation = attributes[:'autoNavigation']
+      end
+
+      if attributes.has_key?(:'autoRespondedReason')
+        self.auto_responded_reason = attributes[:'autoRespondedReason']
       end
 
       if attributes.has_key?(:'bulkRecipientsUri')
@@ -705,6 +722,14 @@ module DocuSign_eSign
         self.phone_authentication = attributes[:'phoneAuthentication']
       end
 
+      if attributes.has_key?(:'phoneNumber')
+        self.phone_number = attributes[:'phoneNumber']
+      end
+
+      if attributes.has_key?(:'phoneNumberMetadata')
+        self.phone_number_metadata = attributes[:'phoneNumberMetadata']
+      end
+
       if attributes.has_key?(:'proofFile')
         self.proof_file = attributes[:'proofFile']
       end
@@ -890,6 +915,7 @@ module DocuSign_eSign
           agent_can_edit_name == o.agent_can_edit_name &&
           allow_system_override_for_locked_recipient == o.allow_system_override_for_locked_recipient &&
           auto_navigation == o.auto_navigation &&
+          auto_responded_reason == o.auto_responded_reason &&
           bulk_recipients_uri == o.bulk_recipients_uri &&
           can_sign_offline == o.can_sign_offline &&
           client_user_id == o.client_user_id &&
@@ -938,6 +964,8 @@ module DocuSign_eSign
           note_metadata == o.note_metadata &&
           offline_attributes == o.offline_attributes &&
           phone_authentication == o.phone_authentication &&
+          phone_number == o.phone_number &&
+          phone_number_metadata == o.phone_number_metadata &&
           proof_file == o.proof_file &&
           recipient_attachments == o.recipient_attachments &&
           recipient_authentication_status == o.recipient_authentication_status &&
@@ -986,7 +1014,7 @@ module DocuSign_eSign
     # Calculates hash code according to all attributes.
     # @return [Fixnum] Hash code
     def hash
-      [access_code, access_code_metadata, add_access_code_to_email, additional_notifications, agent_can_edit_email, agent_can_edit_name, allow_system_override_for_locked_recipient, auto_navigation, bulk_recipients_uri, can_sign_offline, client_user_id, completed_count, creation_reason, custom_fields, declined_date_time, declined_reason, default_recipient, delivered_date_time, delivery_method, delivery_method_metadata, designator_id, designator_id_guid, document_visibility, email, email_metadata, email_notification, embedded_recipient_start_url, error_details, excluded_documents, fax_number, fax_number_metadata, first_name, first_name_metadata, full_name, full_name_metadata, id_check_configuration_name, id_check_configuration_name_metadata, id_check_information_input, identity_verification, inherit_email_notification_configuration, is_bulk_recipient, is_bulk_recipient_metadata, last_name, last_name_metadata, live_oak_start_url, locked_recipient_phone_auth_editable, locked_recipient_sms_editable, name, name_metadata, notary_id, notary_signers, notary_type, note, note_metadata, offline_attributes, phone_authentication, proof_file, recipient_attachments, recipient_authentication_status, recipient_feature_metadata, recipient_id, recipient_id_guid, recipient_signature_providers, recipient_supplies_tabs, recipient_type, recipient_type_metadata, require_id_lookup, require_id_lookup_metadata, require_signer_certificate, require_sign_on_paper, require_upload_signature, role_name, routing_order, routing_order_metadata, sent_date_time, signature_info, signed_date_time, sign_in_each_location, sign_in_each_location_metadata, signing_group_id, signing_group_id_metadata, signing_group_name, signing_group_users, sms_authentication, social_authentications, status, status_code, suppress_emails, tabs, template_locked, template_required, total_tab_count, user_id].hash
+      [access_code, access_code_metadata, add_access_code_to_email, additional_notifications, agent_can_edit_email, agent_can_edit_name, allow_system_override_for_locked_recipient, auto_navigation, auto_responded_reason, bulk_recipients_uri, can_sign_offline, client_user_id, completed_count, creation_reason, custom_fields, declined_date_time, declined_reason, default_recipient, delivered_date_time, delivery_method, delivery_method_metadata, designator_id, designator_id_guid, document_visibility, email, email_metadata, email_notification, embedded_recipient_start_url, error_details, excluded_documents, fax_number, fax_number_metadata, first_name, first_name_metadata, full_name, full_name_metadata, id_check_configuration_name, id_check_configuration_name_metadata, id_check_information_input, identity_verification, inherit_email_notification_configuration, is_bulk_recipient, is_bulk_recipient_metadata, last_name, last_name_metadata, live_oak_start_url, locked_recipient_phone_auth_editable, locked_recipient_sms_editable, name, name_metadata, notary_id, notary_signers, notary_type, note, note_metadata, offline_attributes, phone_authentication, phone_number, phone_number_metadata, proof_file, recipient_attachments, recipient_authentication_status, recipient_feature_metadata, recipient_id, recipient_id_guid, recipient_signature_providers, recipient_supplies_tabs, recipient_type, recipient_type_metadata, require_id_lookup, require_id_lookup_metadata, require_signer_certificate, require_sign_on_paper, require_upload_signature, role_name, routing_order, routing_order_metadata, sent_date_time, signature_info, signed_date_time, sign_in_each_location, sign_in_each_location_metadata, signing_group_id, signing_group_id_metadata, signing_group_name, signing_group_users, sms_authentication, social_authentications, status, status_code, suppress_emails, tabs, template_locked, template_required, total_tab_count, user_id].hash
     end
 
     # Builds the object from hash