+++ lib/docusign_esign/models/signer_attachment.rb
@@ -100,6 +100,9 @@ module DocuSign_eSign
 
     attr_accessor :form_page_number_metadata
 
+    # 
+    attr_accessor :hand_draw_required
+
     # Height of the tab in pixels.
     attr_accessor :height
 
@@ -243,6 +246,7 @@ module DocuSign_eSign
         :'form_page_label_metadata' => :'formPageLabelMetadata',
         :'form_page_number' => :'formPageNumber',
         :'form_page_number_metadata' => :'formPageNumberMetadata',
+        :'hand_draw_required' => :'handDrawRequired',
         :'height' => :'height',
         :'height_metadata' => :'heightMetadata',
         :'merge_field' => :'mergeField',
@@ -326,6 +330,7 @@ module DocuSign_eSign
         :'form_page_label_metadata' => :'PropertyMetadata',
         :'form_page_number' => :'String',
         :'form_page_number_metadata' => :'PropertyMetadata',
+        :'hand_draw_required' => :'String',
         :'height' => :'String',
         :'height_metadata' => :'PropertyMetadata',
         :'merge_field' => :'MergeField',
@@ -517,6 +522,10 @@ module DocuSign_eSign
 
       if attributes.has_key?(:'formPageNumberMetadata')
         self.form_page_number_metadata = attributes[:'formPageNumberMetadata']
+      end
+
+      if attributes.has_key?(:'handDrawRequired')
+        self.hand_draw_required = attributes[:'handDrawRequired']
       end
 
       if attributes.has_key?(:'height')
@@ -743,6 +752,7 @@ module DocuSign_eSign
           form_page_label_metadata == o.form_page_label_metadata &&
           form_page_number == o.form_page_number &&
           form_page_number_metadata == o.form_page_number_metadata &&
+          hand_draw_required == o.hand_draw_required &&
           height == o.height &&
           height_metadata == o.height_metadata &&
           merge_field == o.merge_field &&
@@ -796,7 +806,7 @@ module DocuSign_eSign
     # Calculates hash code according to all attributes.
     # @return [Fixnum] Hash code
     def hash
-      [anchor_allow_white_space_in_characters, anchor_allow_white_space_in_characters_metadata, anchor_case_sensitive, anchor_case_sensitive_metadata, anchor_horizontal_alignment, anchor_horizontal_alignment_metadata, anchor_ignore_if_not_present, anchor_ignore_if_not_present_metadata, anchor_match_whole_word, anchor_match_whole_word_metadata, anchor_string, anchor_string_metadata, anchor_tab_processor_version, anchor_tab_processor_version_metadata, anchor_units, anchor_units_metadata, anchor_x_offset, anchor_x_offset_metadata, anchor_y_offset, anchor_y_offset_metadata, conditional_parent_label, conditional_parent_label_metadata, conditional_parent_value, conditional_parent_value_metadata, custom_tab_id, custom_tab_id_metadata, document_id, document_id_metadata, error_details, form_order, form_order_metadata, form_page_label, form_page_label_metadata, form_page_number, form_page_number_metadata, height, height_metadata, merge_field, merge_field_xml, name, name_metadata, optional, optional_metadata, page_number, page_number_metadata, recipient_id, recipient_id_guid, recipient_id_guid_metadata, recipient_id_metadata, scale_value, scale_value_metadata, smart_contract_information, source, status, status_metadata, tab_group_labels, tab_group_labels_metadata, tab_id, tab_id_metadata, tab_label, tab_label_metadata, tab_order, tab_order_metadata, tab_type, tab_type_metadata, template_locked, template_locked_metadata, template_required, template_required_metadata, tooltip, tool_tip_metadata, width, width_metadata, x_position, x_position_metadata, y_position, y_position_metadata].hash
+      [anchor_allow_white_space_in_characters, anchor_allow_white_space_in_characters_metadata, anchor_case_sensitive, anchor_case_sensitive_metadata, anchor_horizontal_alignment, anchor_horizontal_alignment_metadata, anchor_ignore_if_not_present, anchor_ignore_if_not_present_metadata, anchor_match_whole_word, anchor_match_whole_word_metadata, anchor_string, anchor_string_metadata, anchor_tab_processor_version, anchor_tab_processor_version_metadata, anchor_units, anchor_units_metadata, anchor_x_offset, anchor_x_offset_metadata, anchor_y_offset, anchor_y_offset_metadata, conditional_parent_label, conditional_parent_label_metadata, conditional_parent_value, conditional_parent_value_metadata, custom_tab_id, custom_tab_id_metadata, document_id, document_id_metadata, error_details, form_order, form_order_metadata, form_page_label, form_page_label_metadata, form_page_number, form_page_number_metadata, hand_draw_required, height, height_metadata, merge_field, merge_field_xml, name, name_metadata, optional, optional_metadata, page_number, page_number_metadata, recipient_id, recipient_id_guid, recipient_id_guid_metadata, recipient_id_metadata, scale_value, scale_value_metadata, smart_contract_information, source, status, status_metadata, tab_group_labels, tab_group_labels_metadata, tab_id, tab_id_metadata, tab_label, tab_label_metadata, tab_order, tab_order_metadata, tab_type, tab_type_metadata, template_locked, template_locked_metadata, template_required, template_required_metadata, tooltip, tool_tip_metadata, width, width_metadata, x_position, x_position_metadata, y_position, y_position_metadata].hash
     end
 
     # Builds the object from hash